Collapse into single shared library
Turns out on Windows, each DLL needs to be able to find
liblua, libsodium - so when you build, you wind up with
copies of said libraries all over the place.

Lua supports the "allinone" loader, where if a C module with the root
name is found, it can load from that. So we can just combine everything
into a single module.

I still have a lot of static functions. I do like *only* exporting
functions that users should import, but I'm thinking I should un-static
some of the shared functions. They'll be visible in the shared
object in Unix, but that's not really a big deal.