Constant-loading revamp
One thing I was never a big fan of was how constant values were loaded.

I had macros that assumed they were all size_t types - libsodium has
a few ints and strings as well. So now I track that the value type is.

Additionally when loading via FFI - whether via the "function pointers
in a C module" method or the "strictly FFI" method, I'd prefer for
constants to be loaded in the same manner. So the C module now passes
function pointers for loading the values.